+++ Assets/Scripts/2. Enemies/EnemyHealthManager.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class EnemyHealthManager : MonoBehaviour
 {
-    [SerializeField] private int maxHealth = 1;
-    public int health = 1;
+    [SerializeField] private float maxHealth = 1;
+    public float health = 1;
 
     // Flashing effect variables
     [SerializeField] private float flashDuration = 0.1f;
@@ -24,7 +24,7 @@ public class EnemyHealthManager : MonoBehaviour
         health = maxHealth;
     }
 
-    public void DamageEnemy(int damage)
+    public void DamageEnemy(float damage)
     {
         health -= damage;
         StartCoroutine(FlashEffect());
